backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pytest==8.3.3
pytest-cov==5.0.0
httpx==0.27.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
alembic==1.13.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Logging & Monitoring
structlog==24.1.0
sentry-sdk[fastapi]==2.8.0

# AI Agents SDK
openai-agents==0.13.4
